+++ packages/cx-wealth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+
+### v0.1.1
+
+- 增加了 CxHighlighter 类，用于高亮显示 CX 相关的日志信息
+- 自动安装 CxHighlighter 作为全局输出的高粱显示工具，默认情况下不影响 WealthHelp 的输出
+- 为 rich_types 增加了 rich.box 的默认导入
+- 为 wealth_help 增加了一种 Note 类型，用于在帮助文档中增加一些额外的说明。使用 WealthHelp.add_note() 方法添加 Note 对象。
+
 ### v0.1.0.3
 
 - 修复了 IndexedListPanel 中的计数 bug